[telltale] Move metric rendering into its own module
Keep registry collection, forwarding, status, and routes in telltale.py while moving status-table and histogram rendering into rigging.metrics. Remove the temporary design and research artifacts now that the decision is captured in the pull request.
